Fix : fixed the bugs on OrgMarquee by adding instead constant 3 rows of logo
