+++ frontend/src/pages/LandingPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import Paskilan from '@/assets/Paskilan.png';
 import { Button } from "@/components/ui/button"
-import OrgMarquee from '@/components/layouts/landing_page/Marquee';
+import OrgMarquee from '@/components/layouts/landing_page/OrgMarquee';
 
 
 const LandingPage = () => {
